Add checkins by hour page Prints out the checkins by hour on a page so frequent checkers are not required to download a CSV every time.
