+++ uber/site_sections/reg_reports.py
@@ -1,10 +1,22 @@
+import six
+import calendar
 from collections import defaultdict
 from sqlalchemy import or_, and_
 from sqlalchemy.sql import func
+from sqlalchemy.sql.expression import literal
 
 from uber.config import c
-from uber.decorators import all_renderable, log_pageview, streamable
-from uber.models import Attendee, Group, PromoCode, ReceiptTransaction, ModelReceipt, ReceiptItem
+from uber.decorators import all_renderable, log_pageview, csv_file
+from uber.models import Attendee, Group, PromoCode, ReceiptTransaction, ModelReceipt, ReceiptItem, Tracking
+from uber.utils import localize_datetime
+
+
+def date_trunc_hour(*args, **kwargs):
+    # sqlite doesn't support date_trunc
+    if c.SQLALCHEMY_URL.startswith('sqlite'):
+        return func.strftime(literal('%Y-%m-%d %H:00'), *args, **kwargs)
+    else:
+        return func.date_trunc(literal('hour'), *args, **kwargs)
 
 
 @all_renderable()
@@ -129,3 +141,86 @@ class Root:
             'refund_models': refund_models,
             'counts': counts,
         }
+    
+    def checkins_by_hour(self, session):
+        query_result = session.query(
+            date_trunc_hour(Attendee.checked_in),
+            func.count(date_trunc_hour(Attendee.checked_in))
+        ) \
+            .filter(Attendee.checked_in.isnot(None)) \
+            .group_by(date_trunc_hour(Attendee.checked_in)) \
+            .order_by(date_trunc_hour(Attendee.checked_in))
+
+        hourly_checkins = dict()
+        daily_checkins = defaultdict(int)
+        outside_event_checkins = []
+        for result in query_result:
+            localized_hour = localize_datetime(result[0])
+            hourly_checkins[localized_hour] = result[1]
+            if localized_hour > c.EPOCH and localized_hour < c.ESCHATON:
+                daily_checkins[calendar.day_name[localized_hour.weekday()]] += result[1]
+            else:
+                outside_event_checkins.append(localized_hour)
+
+        return {
+            'checkins': hourly_checkins,
+            'daily_checkins': daily_checkins,
+            'outside_event_checkins': outside_event_checkins,
+        }
+
+    @csv_file
+    def checkins_by_hour_csv(self, out, session):
+        out.writerow(["Time", "# Checked In"])
+        query_result = session.query(
+            date_trunc_hour(Attendee.checked_in),
+            func.count(date_trunc_hour(Attendee.checked_in))
+        ) \
+            .filter(Attendee.checked_in.isnot(None)) \
+            .group_by(date_trunc_hour(Attendee.checked_in)) \
+            .order_by(date_trunc_hour(Attendee.checked_in)) \
+            .all()
+
+        for result in query_result:
+            hour = localize_datetime(result[0])
+            count = result[1]
+            out.writerow([hour, count])
+
+    @csv_file
+    def checkins_by_admin_by_hour(self, out, session):
+        header = ["Time", "Total Checked In"]
+        admins = session.query(Tracking.who).filter(Tracking.action == c.UPDATED,
+                                                    Tracking.model == "Attendee",
+                                                    Tracking.data.contains("checked_in='None -> datetime")
+                                                    ).group_by(Tracking.who).order_by(Tracking.who).distinct().all()
+        for admin in admins:
+            if not isinstance(admin, six.string_types):
+                admin = admin[0]  # SQLAlchemy quirk
+
+            header.append(f"{admin} # Checked In")
+
+        out.writerow(header)
+
+        query_result = session.query(
+            date_trunc_hour(Attendee.checked_in),
+            func.count(date_trunc_hour(Attendee.checked_in))
+        ) \
+            .filter(Attendee.checked_in.isnot(None)) \
+            .group_by(date_trunc_hour(Attendee.checked_in)) \
+            .order_by(date_trunc_hour(Attendee.checked_in)) \
+            .all()
+
+        for result in query_result:
+            hour = localize_datetime(result[0])
+            count = result[1]
+            row = [hour, count]
+            for admin in admins:
+                if not isinstance(admin, six.string_types):
+                    admin = admin[0]  # SQLAlchemy quirk
+                admin_count = session.query(Tracking.which).filter(
+                    date_trunc_hour(Tracking.when) == result[0],
+                    Tracking.who == admin,
+                    Tracking.action == c.UPDATED,
+                    Tracking.model == "Attendee",
+                    Tracking.data.contains("checked_in='None -> datetime")).group_by(Tracking.which).count()
+                row.append(admin_count)
+            out.writerow(row)
